minor tweaks
- got rid of a useless else clause
- swapped the nesting of for loops around in `__consecutive_case_score`
to avoid zipping thrice (this is really minor, but once I noticed it, it
really annoyed me)
- added in a check in `__sequential_letters_score` to make sure we are
counting letters, not digits. It's worth noting that this only checks
against 'a' through 'z', should this be expanded for all lowercase
characters, as it is in the rest of the program? If so, are they always
consecutive with regards to `ord`?